agents/mad_mobility/requirements.txt
google-cloud-aiplatform[agent_engines,adk]
google-adk
requests
